Refactored directory structure to have logical subdirectories
