+++ arduino-main/src/equipment/input/input.cpp
