+++ arduino-main/src/equipment/input/input.h
