+++ arduino-main/src/equipment/input/temperature.cpp
